requirements.txt
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
PyJWT==2.8.0
redis==5.0.3
requests==2.31.0
urllib3==2.2.1
vine==5.1.0
